+++ packages/webpack-cli/lib/utils/arg-parser.js
@@ -2,6 +2,7 @@ const commander = require('commander');
 const logger = require('./logger');
 const { commands } = require('./cli-flags');
 const runHelp = require('../groups/runHelp');
+const runVersion = require('../groups/runVersion');
 const { defaultCommands } = require('./commands');
 
 /**
@@ -13,8 +14,9 @@ const { defaultCommands } = require('./commands');
  * @param {boolean} argsOnly false if all of process.argv has been provided, true if
  * args is only a subset of process.argv that removes the first couple elements
  */
-function argParser(options, args, argsOnly = false, name = '') {
+const argParser = (options, args, argsOnly = false, name = '') => {
     const parser = new commander.Command();
+
     // Set parser name
     parser.name(name);
     parser.storeOptionsAsProperties(false);
@@ -28,19 +30,27 @@ function argParser(options, args, argsOnly = false, name = '') {
             .allowUnknownOption(true)
             .action(async () => {
                 const cliArgs = args.slice(args.indexOf(cmd.name) + 1 || args.indexOf(cmd.alias) + 1);
-                return await require('../commands/resolveCommand').run(defaultCommands[cmd.name], ...cliArgs);
+
+                return await require('./resolve-command')(defaultCommands[cmd.name], ...cliArgs);
             });
+
         return parser;
     }, parser);
 
     // Prevent default behavior
     parser.on('command:*', () => {});
 
-    // Use customized help output if available
-    parser.on('option:help', () => {
+    // Use customized help output
+    if (args.includes('--help') || args.includes('help')) {
         runHelp(args);
         process.exit(0);
-    });
+    }
+
+    // Use Customized version
+    if (args.includes('--version') || args.includes('version') || args.includes('-v')) {
+        runVersion(args);
+        process.exit(0);
+    }
 
     // Allow execution if unknown arguments are present
     parser.allowUnknownOption(true);
@@ -49,6 +59,7 @@ function argParser(options, args, argsOnly = false, name = '') {
     options.reduce((parserInstance, option) => {
         let optionType = option.type;
         let isStringOrBool = false;
+
         if (Array.isArray(optionType)) {
             // filter out duplicate types
             optionType = optionType.filter((type, index) => {
@@ -75,7 +86,9 @@ function argParser(options, args, argsOnly = false, name = '') {
         }
 
         const flags = option.alias ? `-${option.alias}, --${option.name}` : `--${option.name}`;
+
         let flagsWithType = flags;
+
         if (isStringOrBool) {
             // commander recognizes [value] as an optional placeholder,
             // making this flag work either as a string or a boolean
@@ -89,6 +102,36 @@ function argParser(options, args, argsOnly = false, name = '') {
             if (option.multiple) {
                 // a multiple argument parsing function
                 const multiArg = (value, previous = []) => previous.concat([value]);
+                parserInstance.option(flagsWithType, option.description, multiArg, option.defaultValue).action(() => {});
+            } else if (option.multipleType) {
+                // for options which accept multiple types like env
+                // so you can do `--env platform=staging --env production`
+                // { platform: "staging", production: true }
+                const multiArg = (value, previous = {}) => {
+                    // this ensures we're only splitting by the first `=`
+                    const [allKeys, val] = value.split(/=(.+)/, 2);
+                    const splitKeys = allKeys.split(/\.(?!$)/);
+
+                    let prevRef = previous;
+
+                    splitKeys.forEach((someKey, index) => {
+                        if (!prevRef[someKey]) {
+                            prevRef[someKey] = {};
+                        }
+
+                        if ('string' === typeof prevRef[someKey]) {
+                            prevRef[someKey] = {};
+                        }
+
+                        if (index === splitKeys.length - 1) {
+                            prevRef[someKey] = val || true;
+                        }
+
+                        prevRef = prevRef[someKey];
+                    });
+
+                    return previous;
+                };
                 parserInstance.option(flagsWithType, option.description, multiArg, option.defaultValue).action(() => {});
             } else {
                 // Prevent default behavior for standalone options
@@ -120,8 +163,8 @@ function argParser(options, args, argsOnly = false, name = '') {
 
     const result = parser.parse(args, parseOptions);
     const opts = result.opts();
-
     const unknownArgs = result.args;
+
     args.forEach((arg) => {
         const flagName = arg.slice(5);
         const option = options.find((opt) => opt.name === flagName);
@@ -129,6 +172,7 @@ function argParser(options, args, argsOnly = false, name = '') {
         const flagUsed = args.includes(flag) && !unknownArgs.includes(flag);
         let alias = '';
         let aliasUsed = false;
+
         if (option && option.alias) {
             alias = `-${option.alias}`;
             aliasUsed = args.includes(alias) && !unknownArgs.includes(alias);
@@ -155,6 +199,6 @@ function argParser(options, args, argsOnly = false, name = '') {
         unknownArgs,
         opts,
     };
-}
+};
 
 module.exports = argParser;