+++ test/analyze/main.js
@@ -1,1 +1,0 @@
-console.log('webpack-bundle-analyzer is installed');