+++ test/bail/no-bail-webpack.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+    entry: './src/first.js',
+    mode: 'development',
+};