+++ test/config-format/typescript/typescript.test.js
@@ -1,10 +1,7 @@
-/* eslint-disable   node/no-missing-require */
 /* eslint-disable   node/no-unpublished-require */
 const { run, runInstall } = require('../../utils/test-utils');
 const { stat } = require('fs');
 const { resolve } = require('path');
-
-jest.setTimeout(1000 * 60 * 5);
 
 describe('webpack cli', () => {
     it(