+++ test/config-name/config-name.test.js
@@ -6,11 +6,12 @@ const { resolve } = require('path');
 
 describe('--config-name flag', () => {
     it('should select only the config whose name is passed with --config-name', (done) => {
-        const { stderr, stdout } = run(__dirname, ['--config-name', 'first'], false);
+        const { stderr, stdout, exitCode } = run(__dirname, ['--config-name', 'first'], false);
         expect(stderr).toBeFalsy();
         expect(stdout).toContain('first');
         expect(stdout).not.toContain('second');
         expect(stdout).not.toContain('third');
+        expect(exitCode).toBe(0);
 
         stat(resolve(__dirname, './dist/dist-first.js'), (err, stats) => {
             expect(err).toBe(null);
@@ -20,33 +21,64 @@ describe('--config-name flag', () => {
     });
 
     it('should work with multiple values for --config-name', (done) => {
-        const { stderr, stdout } = run(__dirname, ['--config-name', 'first', '--config-name', 'third'], false);
+        const { stderr, stdout, exitCode } = run(__dirname, ['--config-name', 'first', '--config-name', 'third'], false);
         expect(stderr).toBeFalsy();
         expect(stdout).toContain('first');
         expect(stdout).not.toContain('second');
         expect(stdout).toContain('third');
+        expect(exitCode).toBe(0);
 
-        stat(resolve(__dirname, './dist/dist-first.js'), (err, stats) => {
-            expect(err).toBe(null);
-            expect(stats.isFile()).toBe(true);
-            done();
-        });
         stat(resolve(__dirname, './dist/dist-third.js'), (err, stats) => {
             expect(err).toBe(null);
             expect(stats.isFile()).toBe(true);
-            done();
+
+            stat(resolve(__dirname, './dist/dist-first.js'), (err, stats) => {
+                expect(err).toBe(null);
+                expect(stats.isFile()).toBe(true);
+                done();
+            });
         });
     });
 
     it('should log error if invalid config name is provided', () => {
-        const { stderr, stdout } = run(__dirname, ['--config-name', 'test'], false);
+        const { stderr, stdout, exitCode } = run(__dirname, ['--config-name', 'test'], false);
+
         expect(stderr).toContain('Configuration with name "test" was not found.');
         expect(stdout).toBeFalsy();
+        expect(exitCode).toBe(2);
     });
 
     it('should log error if multiple configurations are not found', () => {
-        const { stderr, stdout } = run(__dirname, ['--config-name', 'test', '-c', 'single-config.js'], false);
-        expect(stderr).toContain('Multiple configurations not found. Please use "--config-name" with multiple configurations');
+        const { stderr, stdout, exitCode } = run(__dirname, ['--config-name', 'test', '-c', 'single-config.js'], false);
+
+        expect(stderr).toContain('Configuration with name "test" was not found.');
         expect(stdout).toBeFalsy();
+        expect(exitCode).toBe(2);
+    });
+
+    it('should log error if multiple configurations are not found #1', () => {
+        const { stderr, stdout, exitCode } = run(
+            __dirname,
+            ['--config-name', 'test', '--config-name', 'bar', '-c', 'single-config.js'],
+            false,
+        );
+
+        expect(stderr).toContain('Configuration with name "test" was not found.');
+        expect(stderr).toContain('Configuration with name "bar" was not found.');
+        expect(stdout).toBeFalsy();
+        expect(exitCode).toBe(2);
+    });
+
+    it('should log error if multiple configurations are not found #2', () => {
+        const { stderr, stdout, exitCode } = run(
+            __dirname,
+            ['--config-name', 'first', '--config-name', 'bar', '-c', 'single-config.js'],
+            false,
+        );
+
+        expect(stderr).not.toContain('Configuration with name "first" was not found.');
+        expect(stderr).toContain('Configuration with name "bar" was not found.');
+        expect(stdout).toBeFalsy();
+        expect(exitCode).toBe(2);
     });
 });