+++ packages/webpack-scaffold/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.2](https://github.com/webpack/webpack-cli/compare/@webpack-cli/webpack-scaffold@1.0.1...@webpack-cli/webpack-scaffold@1.0.2) (2020-10-19)
+
+### Bug Fixes
+
+-   support array of functions and promises ([#1946](https://github.com/webpack/webpack-cli/issues/1946)) ([2ace39b](https://github.com/webpack/webpack-cli/commit/2ace39b06117f558c0d8528cea9248253cbdf593))
+
+## [1.0.1](https://github.com/webpack/webpack-cli/compare/@webpack-cli/webpack-scaffold@1.0.1-rc.1...@webpack-cli/webpack-scaffold@1.0.1) (2020-10-10)
+
+### Bug Fixes
+
+-   upgrade lock file ([#1885](https://github.com/webpack/webpack-cli/issues/1885)) ([8df291e](https://github.com/webpack/webpack-cli/commit/8df291eef0fad7c91d912b158b3c2915cddfacd1))
+
 ## [1.0.1-rc.1](https://github.com/webpack/webpack-cli/compare/@webpack-cli/webpack-scaffold@1.0.1-alpha.4...@webpack-cli/webpack-scaffold@1.0.1-rc.1) (2020-10-06)
 
 **Note:** Version bump only for package @webpack-cli/webpack-scaffold